client/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vtms-client"
version = "0.1.0"
description = "VTMS vehicle telemetry client (OBD-II + GPS → MQTT)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "obd",
    "paho-mqtt",
    "pynmea2",
    "pyserial",
    "pygeohash",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
]

[project.scripts]
vtms-client = "vtms_client.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
python-preference = "system"
default-groups = ["dev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers --disable-warnings --color=yes"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
